13_julia_fp: send location in float, rather than int
This avoids an int-to-float cast that is currently unimplemented

+++ test/unit_tests/13_julia_fp/kernel.cpp
@@ -92,4 +92,4 @@ void julia(unsigned *framebuffer, float cr, float ci)
   }
 }
 
-void func(int32_t *a, int32_t *b) { julia((unsigned *)b, a[0], a[1]); }
+void func(float *a, float *b) { julia((unsigned *)b, a[0], a[1]); }
